Extract WidgetEditor to its own file (including moving SectionControlButton into components/)
